requirements.txt
openai
python-dotenv
gspread
oauth2client
